mars, 2020: Add an error component that shows how to handle the 404 and other errors
