+++ .changeset/large-flowers-occur.md
@@ -1,0 +1,6 @@
+---
+"@frontity/mars-theme": patch
+"@frontity/twentytwenty-theme": patch
+---
+
+Add a new component that shows how to handle `data.isError` in mars-theme and twenty-twenty theme
